+++ Group-Assignment/app/page.js
@@ -1,14 +1,12 @@
-import RegistrationForm from "./components/RegistrationForm";
-import StudentList from "./components/StudentList";
-import CombinedComponent from "./components/CombindedComponent";
+import StudentForm from "./components/StudentForm";
 
 
 
 export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center p-8">
-    <h1 className="text-4xl font-bold">Welcome to my App</h1>
-   <CombinedComponent/>
+    <h1 className="text-4xl font-bold">Change to School name</h1>
+   <StudentForm/>
     </main>
   );
 }
